rename fromPairs to fromEntries
